refactor(core): isolate prompt materialization
